requirements.txt
psycopg2-binary~=2.9.10
pandas~=2.2.3
databento~=0.50.0
SQLAlchemy~=2.0.39
streamlit~=1.43.2
plotly~=6.0.1
PyYAML~=6.0.2
dotenv~=0.9.9
python-dotenv~=1.1.0
numpy~=2.2.4